% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_abc_means}
\alias{simple_abc_means}
\title{Simple example: group means}
\format{
A data frame with 5 rows and 2 columns: \code{group} (character labels
\code{"1"} to \code{"5"}) and \code{mean} (numeric).
}
\source{
Ennis, J. M., Fayle, C. M., and Ennis, D. M. (2012). Assignment-minimum
clique coverings. \emph{ACM Journal of Experimental Algorithmics}, 17, Article
1.5. \doi{10.1145/2133803.2275596}
}
\usage{
simple_abc_means
}
\description{
The group means that go with \link{simple_abc_pairs}. They fix the order of the
groups and of the letters.
}
\examples{
simple_abc_means
}
\seealso{
\link{simple_abc_pairs}, \code{\link[=reduce_letters]{reduce_letters()}}
}
\keyword{datasets}
